tools/pyproject.toml
[tool.poetry]
name = "plugins"
version = "0.1.0"
description = ""
authors = ["michael1011 <me@michael1011.at>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-bitcoinrpc = "^1.0"
pyotp = "^2.8.0"
sseclient = "^0.0.27"
ruff = "^0.1.3"
pyln-client = "^23.5"
webdavclient3 = "^3.14.6"
pytest = "^7.4.3"
bolt11 = "^2.0.1"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
grpc-interceptor = "^0.15.2"
cryptography = "^41.0.5"
psycopg = {extras = ["binary"], version = "^3.1.12"}
sqlalchemy = "^2.0.22"
cachetools = "^5.3.2"
pandas = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["ALL"]
exclude = ["plugins/hold/protos"]
line-length = 100
ignore = [
    "T201", "D101", "D211", "D213", "INP001", "BLE001", "FBT001", "FBT002", "FBT003", "S605",
    "TD002", "TD003", "FIX002", "ANN101", "D102", "D103", "D107", "D100", "SLOT000", "S101",
    "PLR2004", "ARG001", "PLR0913", "D104", "FA102", "COM812", "UP032", "ISC001", "A003"
]

[tool.ruff.format]
quote-style = "double"
